+++ js/get-random-float.js
@@ -1,0 +1,12 @@
+
+
+function getRandomFloating(min, max, numberDecimal) {
+  if ((min < 0) || (max < 0) || (max <= min)) {
+    throw new Error('Неверное значение');
+  } else {
+    const randFloat = (min + Math.random() * (max - min));
+    return randFloat.toFixed(numberDecimal);
+  }
+}
+
+export {getRandomFloating};
